requirements.txt
ipykernel==7.1.0
matplotlib==3.10.0
pandas == 3.0.0
torchinfo == 1.8.0
torchmetrics == 1.8.2
tqdm == 4.67.3
kornia == 0.8.2
brevitas == 0.12.1
netron == 8.9.9
onnx == 1.21.0
onnxscript == 0.6.2
onnxoptimizer == 0.4.2
#ipywidgets == 8.0.0